fix: http server request size (#7609)
Description
---
Fixed http server request size. When byte size request is converted to
JSON, it expands ~2.3 times. This PR increase the http server size by
2.5 times 4MiB.

This is not strictly a breaking change as the server will not reject
previously accepted transactions; it will now accept previously rejected
trnsactions.

Motivation and Context
---
Transactions of 2MiB size could not be submitted to the base node, as
they were rejected with
```
[wallet::transaction_service::protocols] DEBUG Transaction '12820213401082185863' size ok, can be broadcast (got: 2109809, limit: 4184064).
```
```
 [tari::wallet::client::http] [Thread:4504] WARN  Received error response from Base Node wallet service: 413 Payload Too Large. length limit exceeded
 ```

How Has This Been Tested?
---
System-level testing.

This transaction was previously in the Completed state forever, now mined.

<img width="1136" height="209" alt="image" src="https://github.com/user-attachments/assets/65705020-6536-4dc9-869d-d4ee627b2533" />


What process can a PR reviewer use to test or verify this change?
---
Code review.
System-level testing.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a database, resync the chain -->